pyproject.toml
[tool.poetry]
name = "driconfig"
version = "0.1.0"
description = "Pydantic-ish YAML configuration management."
license = "MIT"
authors = ["Dribia Data Research <info@dribia.com>"]
maintainers = [
    "Irene Pérez <irene@dribia.com>",
    "Albert Iribarne <iribarne@dribia.com>",
]
readme = "README.md"
homepage = "https://dribia.github.io/driconfig"
repository = "https://github.com/dribia/driconfig"
documentation = "https://dribia.github.io/driconfig"

[tool.poetry.dependencies]
python = ">=3.6.2,<4.0"
importlib-metadata = {version = ">=1.0, <5.0", python = "<3.8"}
pydantic = "^1.8.2"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.15.0"
black = "21.9b0"
isort = "5.9.3"
mypy = "0.910"
pydocstyle = "6.1.1"
flake8 = "3.9.2"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
pytest-mock = "^3.6.1"
mkdocs = "^1.2.2"
mkdocs-material = "^7.3.0"
toml = "^0.10.2"
types-PyYAML = "^5.4.3"

[tool.pydocstyle]
convention = "google"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88
ensure_newline_before_comments = true
known_third_party = ""

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
